fix not entity found exception in delete request, and changed requestMapping in boardrooms

+++ src/main/java/com/krzysztof/studio/boardroom/BoardroomService.java
@@ -37,7 +37,7 @@ public class BoardroomService {
     }
 
     public void delete(String name) {
-        boardroomRepository.deleteById(name);
+        if (boardroomRepository.existsById(name)) boardroomRepository.deleteById(name);
     }
 
     public void update(String name, Boardroom boardroomUpdated) {
